Refactored: created custom isMounted hook, imported where needed
