+++ src/components/cards/UserCard.tsx
@@ -6,7 +6,7 @@ import {
   Button,
 } from "@mui/material";
 import { clientApi } from "../../services/clientApi";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { setUserState } from "../../store/userSlice";
@@ -14,6 +14,7 @@ import { RootState } from "../../store";
 import { clearError, setError } from "../../store/errorSlice";
 import LoadingCard from "./LoadingCard";
 import { User } from "./types/User.types";
+import useMount from "../../services/isMountedHook";
 
 const userInit = {
   user: {
@@ -38,14 +39,7 @@ function UserCard() {
   const { id } = useParams();
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const isMounted = useRef(false);
-
-  useEffect(() => {
-    isMounted.current = true;
-    return () => {
-      isMounted.current = false;
-    };
-  }, []);
+  const isMounted = useMount();
 
   useEffect(() => {
     clientApi.user
@@ -64,7 +58,7 @@ function UserCard() {
         return Promise.reject(err);
       })
       .then((res) => {
-        if (isMounted.current) {
+        if (isMounted()) {
           dispatch(clearError());
           setCurrentUser(res.data);
           setIsLoaded(true);
