+++ src/components/cards/WarehouseConsignmentCard.tsx
@@ -5,7 +5,7 @@ import {
   Button,
   Typography,
 } from "@mui/material";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../store";
@@ -14,6 +14,7 @@ import { clientApi } from "../../services/clientApi";
 import { Alert } from "@mui/material";
 import LoadingCard from "./LoadingCard";
 import { Consignment, UserInfo } from "./types/WarehouseConsignmnet.types";
+import useMount from "../../services/isMountedHook";
 
 const consignmentInit = {
   id: 0,
@@ -49,14 +50,7 @@ export default function WarehouseConsignmentCard() {
   const { id } = useParams();
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const isMounted = useRef(false);
-
-  useEffect(() => {
-    isMounted.current = true;
-    return () => {
-      isMounted.current = false;
-    };
-  }, []);
+  const isMounted = useMount();
 
   useEffect(() => {
     clientApi.consignment
@@ -75,7 +69,7 @@ export default function WarehouseConsignmentCard() {
         return Promise.reject(err);
       })
       .then((res) => {
-        if (isMounted.current) {
+        if (isMounted()) {
           setConsignment(res.data.consignment);
           setUserActions(res.data.actions);
           console.log("consignment", res.data.consignment);
