+++ src/services/clientApi.ts
@@ -13,9 +13,9 @@ import {
   report,
 } from "./clientApi.types";
 
-const baseUrl: string = process.env.REACT_APP_WAREHOUSE_URL as string;
+// const baseUrl: string = process.env.REACT_APP_WAREHOUSE_URL as string;
 //TODO Test local url
-// const baseUrl: string = process.env.REACT_APP_WAREHOUSE_LOCAL_URL as string;
+const baseUrl: string = process.env.REACT_APP_WAREHOUSE_LOCAL_URL as string;
 
 function initClientApi() {
   return {
